+++ mocks/operator/redisfailover/service/RedisFailoverClient.go
@@ -14,6 +14,20 @@ type RedisFailoverClient struct {
 	mock.Mock
 }
 
+// DestroyHaproxyMasterResources provides a mock function with given fields: rFailover
+func (_m *RedisFailoverClient) DestroyHaproxyMasterResources(rFailover *v1.RedisFailover) error {
+	ret := _m.Called(rFailover)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(*v1.RedisFailover) error); ok {
+		r0 = rf(rFailover)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
 // DestroyOrphanedRedisSlaveHaProxy provides a mock function with given fields: rFailover
 func (_m *RedisFailoverClient) DestroyOrphanedRedisSlaveHaProxy(rFailover *v1.RedisFailover) error {
 	ret := _m.Called(rFailover)